Moved authorID to constants file so I don't have to load axiosCalls.ts in Cypress.
