+++ InstaTonne-FrontEnd/src/components/CreatePostPage.vue
@@ -30,7 +30,8 @@
 import { ref } from "vue";
 import PostEditor from "./PostEditor.vue";
 import ErrorPage from "./ErrorPage.vue";
-import { createHTTP, USER_AUTHOR_ID_COOKIE } from "../axiosCalls";
+import { createHTTP } from "../axiosCalls";
+import { USER_AUTHOR_ID_COOKIE } from "../constants";
 import Cookies from "js-cookie";
 
 const authorId = Cookies.get(USER_AUTHOR_ID_COOKIE);
