+++ InstaTonne-FrontEnd/src/constants.ts
@@ -1,0 +1,3 @@
+const USER_AUTHOR_ID_COOKIE = "InstatonneAuthorId";
+
+export { USER_AUTHOR_ID_COOKIE }
